+++ src/Ki/Context.hs
@@ -2,17 +2,20 @@ module Ki.Context
   ( Context (..),
     dummyContext,
     globalContext,
+    newContext,
+    contextCancelTokenSTM,
   )
 where
 
 import Ki.CancelToken (CancelToken)
-import Ki.Ctx (Ctx)
+import Ki.Ctx (CancelState, Ctx)
 import qualified Ki.Ctx as Ctx
 import Ki.Prelude
 
 data Context = Context
   { cancelContext :: IO (),
-    contextCancelTokenSTM :: STM CancelToken,
+    -- | Get this content's current cancel state. This action never retries.
+    contextCancelStateSTM :: STM CancelState,
     -- | Derive a child context from a parent context.
     --
     --   * If the parent is already cancelled, so is the child.
@@ -27,7 +30,7 @@ newContext :: Ctx -> Context
 newContext ctx =
   Context
     { cancelContext = Ctx.cancelCtx ctx,
-      contextCancelTokenSTM = Ctx.ctxCancelToken ctx,
+      contextCancelStateSTM = readTVar (Ctx.cancelStateVar ctx),
       deriveContext = newContext <$> Ctx.deriveCtx ctx
     }
 
@@ -35,7 +38,7 @@ dummyContext :: Context
 dummyContext =
   Context
     { cancelContext = pure (),
-      contextCancelTokenSTM = retry,
+      contextCancelStateSTM = retry,
       deriveContext = pure dummyContext
     }
 
@@ -44,6 +47,12 @@ globalContext :: Context
 globalContext =
   Context
     { cancelContext = pure (),
-      contextCancelTokenSTM = retry,
+      contextCancelStateSTM = pure Ctx.CancelState'NotCancelled,
       deriveContext = newContext <$> Ctx.newCtxSTM
     }
+
+contextCancelTokenSTM :: Context -> STM CancelToken
+contextCancelTokenSTM Context {contextCancelStateSTM} =
+  contextCancelStateSTM >>= \case
+    Ctx.CancelState'NotCancelled -> retry
+    Ctx.CancelState'Cancelled token _way -> pure token