+++ src/Ki/Ctx.hs
@@ -3,11 +3,12 @@
 
 module Ki.Ctx
   ( Ctx (..),
+    CancelState (..),
+    CancelWay (..),
     newCtxSTM,
     deriveCtx,
     cancelCtx,
     cancelCtxSTM,
-    ctxCancelToken,
   )
 where
 
@@ -16,7 +17,7 @@ import Ki.CancelToken
 import Ki.Prelude
 
 data Ctx = Ctx
-  { cancelTokenVar :: TVar (Maybe CancelToken),
+  { cancelStateVar :: TVar CancelState,
     childrenVar :: TVar (IntMap Ctx),
     -- | The next id to assign to a child context. The child needs a unique identifier so it can delete itself from its
     -- parent's children map if it's cancelled independently. Wrap-around seems ok; that's a *lot* of children for one
@@ -34,28 +35,40 @@ data Ctx = Ctx
     onCancel :: STM ()
   }
 
-newCtxSTM :: STM Ctx
-newCtxSTM =
-  newCtxSTM_ (pure ())
+data CancelState
+  = CancelState'NotCancelled
+  | CancelState'Cancelled CancelToken CancelWay
 
-newCtxSTM_ :: STM () -> STM Ctx
-newCtxSTM_ onCancel = do
-  cancelTokenVar <- newTVar Nothing
+data CancelWay
+  = CancelWay'Indirect -- parent (or grandparent, etc...) was cancelled
+  | CancelWay'Direct -- we were cancelled
+
+newCtxSTM :: STM Ctx
+newCtxSTM = do
+  cancelStateVar <- newTVar CancelState'NotCancelled
+  newCtxSTM_ cancelStateVar (pure ())
+
+newCtxSTM_ :: TVar CancelState -> STM () -> STM Ctx
+newCtxSTM_ cancelStateVar onCancel = do
   childrenVar <- newTVar IntMap.empty
   nextIdVar <- newTVar 0
-  pure Ctx {cancelTokenVar, childrenVar, nextIdVar, onCancel}
+  pure Ctx {cancelStateVar, childrenVar, nextIdVar, onCancel}
 
 deriveCtx :: Ctx -> STM Ctx
-deriveCtx context@Ctx {cancelTokenVar, childrenVar, nextIdVar} =
-  readTVar cancelTokenVar >>= \case
-    Nothing -> do
-      childId <- readTVar nextIdVar
-      writeTVar nextIdVar $! childId + 1
-      child <- newCtxSTM_ (modifyTVar' childrenVar (IntMap.delete childId))
-      children <- readTVar childrenVar
-      writeTVar childrenVar $! IntMap.insert childId child children
-      pure child
-    Just (CancelToken _) -> pure context
+deriveCtx Ctx {cancelStateVar, childrenVar, nextIdVar} = do
+  childId <- readTVar nextIdVar
+  writeTVar nextIdVar $! childId + 1
+  child <- do
+    derivedCancelStateVar <- do
+      derivedCancelState <-
+        readTVar cancelStateVar <&> \case
+          CancelState'NotCancelled -> CancelState'NotCancelled
+          CancelState'Cancelled token _cancelWay -> CancelState'Cancelled token CancelWay'Indirect
+      newTVar derivedCancelState
+    newCtxSTM_ derivedCancelStateVar (modifyTVar' childrenVar (IntMap.delete childId))
+  children <- readTVar childrenVar
+  writeTVar childrenVar $! IntMap.insert childId child children
+  pure child
 
 cancelCtx :: Ctx -> IO ()
 cancelCtx context = do
@@ -63,30 +76,17 @@ cancelCtx context = do
   atomically (cancelCtxSTM context token)
 
 cancelCtxSTM :: Ctx -> CancelToken -> STM ()
-cancelCtxSTM ctx@Ctx {onCancel} token =
-  whenCanceling ctx token do
-    cancelChildren ctx token
-    onCancel
-
-ctxCancelSTM_ :: CancelToken -> Ctx -> STM ()
-ctxCancelSTM_ token ctx =
-  whenCanceling ctx token (cancelChildren ctx token)
-
-whenCanceling :: Ctx -> CancelToken -> STM () -> STM ()
-whenCanceling Ctx {cancelTokenVar} token action =
-  readTVar cancelTokenVar >>= \case
-    Nothing -> do
-      writeTVar cancelTokenVar $! Just token
-      action
-    Just (CancelToken _) -> pure ()
-
-cancelChildren :: Ctx -> CancelToken -> STM ()
-cancelChildren Ctx {childrenVar} token = do
-  children <- readTVar childrenVar
-  for_ (IntMap.elems children) (ctxCancelSTM_ token)
-
-ctxCancelToken :: Ctx -> STM CancelToken
-ctxCancelToken Ctx {cancelTokenVar} =
-  readTVar cancelTokenVar >>= \case
-    Nothing -> retry
-    Just token -> pure token
+cancelCtxSTM ctx token =
+  readTVar (cancelStateVar ctx) >>= \case
+    CancelState'NotCancelled -> do
+      cancelChildren ctx
+      writeTVar (cancelStateVar ctx) $! CancelState'Cancelled token CancelWay'Direct
+      onCancel ctx
+    CancelState'Cancelled _token _way -> pure ()
+  where
+    cancelChildren :: Ctx -> STM ()
+    cancelChildren Ctx {childrenVar} = do
+      children <- readTVar childrenVar
+      for_ (IntMap.elems children) \child -> do
+        cancelChildren child
+        writeTVar (cancelStateVar child) $! CancelState'Cancelled token CancelWay'Indirect